fire_refactored/pyproject.toml
[tool.poetry]
name = "fire-simulator"
version = "0.1.0"
description = "A FIRE simulator including total personal finance with Streamlit UI"
authors = ["Moejay10 <mohe95@hotmail.no>"]
readme = "README.md"
packages = [{include = "finance"}]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
pandas = "^2.2.2"
numpy = "^2.1.0"
plotly = "^5.24.0"
openpyxl = "^3.1.5"
pyyaml = "^6.0.2"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
